solution/requirements.txt
fastapi
fastapi[standard]
pillow